+++ Models/Dto/TwoCitiesInfoDto.cs
@@ -1,13 +1,12 @@
-﻿namespace GeoInfo.Models.Dto
+﻿namespace GeoInfo.Models.Dto;
+
+public class TwoCitiesInfoDto
 {
-    public class TwoCitiesInfoDto
+    public TwoCitiesInfoDto(IEnumerable<City> data, string info)
     {
-        public TwoCitiesInfoDto(IEnumerable<City> data, string info)
-        {
-            Data = data;
-            Info = info;
-        }
-        public IEnumerable<City> Data { get; set; }
-        public string Info { get; set; }
+        Data = data;
+        Info = info;
     }
+    public IEnumerable<City> Data { get; set; }
+    public string Info { get; set; }
 }
