+++ Models/NotFoundException.cs
@@ -1,8 +1,7 @@
-﻿namespace GeoInfo.Models
+﻿namespace GeoInfo.Models;
+
+public class NotFoundException : Exception
 {
-    public class NotFoundException : Exception
-    {
-        public NotFoundException(string message)
-        : base(message) { }
-    }
+    public NotFoundException(string message)
+    : base(message) { }
 }
